chore(provider): refactor package layout
Move rest of the resouce and data source related code under their own
packages: cloud, firewall, ip, network, objectstorage, router, tag
