+++ upcloud/resource_upcloud_firewall_rules_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/request"
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/service"
-	"github.com/hashicorp/go-cty/cty"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
@@ -335,27 +334,4 @@ func testUpcloudFirewallRulesInstanceConfigUpdate() string {
 		  }
 
 		}`
-}
-
-func TestFirewallRuleValidateOptionalPort(t *testing.T) {
-	p := cty.Path{}
-	if diag := firewallRuleValidateOptionalPort("1", p); len(diag) > 0 {
-		t.Error(diag[0].Detail)
-	}
-
-	if diag := firewallRuleValidateOptionalPort("65535", p); len(diag) > 0 {
-		t.Error(diag[0].Detail)
-	}
-
-	if diag := firewallRuleValidateOptionalPort("abc", p); len(diag) < 1 {
-		t.Error("firewallRuleValidateOptionalPort failed 'abc' is not valid port")
-	}
-
-	if diag := firewallRuleValidateOptionalPort("0", p); len(diag) < 1 {
-		t.Error("firewallRuleValidateOptionalPort failed '0' is not valid port")
-	}
-
-	if diag := firewallRuleValidateOptionalPort("65536", p); len(diag) < 1 {
-		t.Error("firewallRuleValidateOptionalPort failed '65536' is not valid port")
-	}
 }
